Module debris_common::file_provider

This module has the FileProvider trait, which is used by the compile config to resolve import statements.

Structs

FsFileProvider

A FileProvider implementation that reads files from the file system

MemoryFileProvider

A FileProvider implementation that has an in-memory pool of files

Traits

FileProvider