Module debris_common::graph

Very basic graph implementation which is specialized for llir block analysis

Structs

GraphDfs

A struct that implements a non-recursive Depth-first-search on a GraphMatrix. This is a struct in order to save on allocations in hot code. The elements are yielded in post-order, i.e. every node gets only yielded if all of its children are yielded.

GraphLoopDetector

Detects loops in a graph. This is a custom struct in order to save on allocations in hot code.

GraphMatrix

A directed graph backed by a matrix. T is the type of weights and defaults to (). Since this struct is only intended to be used for the specific use case of control flow analysis a lot of functionality, like resizing, is not implemented.