Module debris_common::input_file

Input files are used to store the files required for successfull compilation The big advantage of have a global storage of all possible input files is that this allows for extremely cheap spans which only consist of a (start, length) tuple. To get the corresponding file for a span, simply check whether the span lies withing file.offset and file.offset + file.length

Structs

Code
A code object contains the source code and optionally a path to the corresponding file

CodeRef
A handy object to work in a specific input file

InputFile
A single input file, implementation detail

InputFiles
Keeps track of all input files and allows to make cheap copy-able spans

Type Definitions

CodeId
The type of a code id, currently just a usize