Crate debris_common

Common items used by the compiler

Macros

accessor

Structs

Code
A code object contains the source code and optionally a path to the corresponding file

LocalSpan
A Span which specifies a unique span of characters in a local file

Span
A span which uniquely specifies a span of characters in their corresponsing file

Enums

Accessor
An accessor. Currently a mess

Ident
identifier for a value

InputFile
Marks any Input file. Will change or be deleted

SpecialIdent
Every special identifier

Type Definitions

CodeRef
A reference to a code object